add LicenseActivity
